+++ tests/framework/population/test_manager.py
@@ -2,25 +2,29 @@ import pytest
 
 from vivarium import Component
 from vivarium.framework.population.exceptions import PopulationError
-from vivarium.framework.population.manager import InitializerComponentSet, PopulationManager
+from vivarium.framework.population.manager import (
+    InitializerComponentSet,
+    PopulationManager,
+    SimulantData,
+)
 
 
-def test_initializer_set_fail_type():
+def test_initializer_set_fail_type() -> None:
     component_set = InitializerComponentSet()
 
     with pytest.raises(TypeError):
-        component_set.add(lambda: "test", ["test_column"])
+        component_set.add(lambda _: None, ["test_column"])
 
-    def initializer():
-        return "test"
+    def initializer(simulant_data: SimulantData) -> None:
+        pass
 
     with pytest.raises(TypeError):
         component_set.add(initializer, ["test_column"])
 
 
 class NonComponent:
-    def initializer(self) -> str:
-        return "test"
+    def initializer(self, simulant_data: SimulantData) -> None:
+        pass
 
 
 class InitializingComponent(Component):
@@ -28,25 +32,25 @@ class InitializingComponent(Component):
     def name(self) -> str:
         return self._name
 
-    def __init__(self, name: str):
+    def __init__(self, name: str) -> None:
         super().__init__()
         self._name = name
 
-    def initializer(self) -> str:
-        return "test"
+    def initializer(self, simulant_data: SimulantData) -> None:
+        pass
 
-    def other_initializer(self) -> str:
-        return "whoops"
+    def other_initializer(self, simulant_data: SimulantData) -> None:
+        pass
 
 
-def test_initializer_set_fail_attr():
+def test_initializer_set_fail_attr() -> None:
     component_set = InitializerComponentSet()
 
     with pytest.raises(AttributeError):
         component_set.add(NonComponent().initializer, ["test_column"])
 
 
-def test_initializer_set_duplicate_component():
+def test_initializer_set_duplicate_component() -> None:
     component_set = InitializerComponentSet()
     component = InitializingComponent("test")
 
@@ -55,7 +59,7 @@ def test_initializer_set_duplicate_component():
         component_set.add(component.other_initializer, ["test_column2"])
 
 
-def test_initializer_set_duplicate_columns():
+def test_initializer_set_duplicate_columns() -> None:
     component_set = InitializerComponentSet()
     component1 = InitializingComponent("test1")
     component2 = InitializingComponent("test2")
@@ -69,14 +73,14 @@ def test_initializer_set_duplicate_columns():
         component_set.add(component2.initializer, ["sneaky_column"] + columns)
 
 
-def test_initializer_set_population_manager():
+def test_initializer_set_population_manager() -> None:
     component_set = InitializerComponentSet()
     population_manager = PopulationManager()
 
     component_set.add(population_manager.on_initialize_simulants, ["tracked"])
 
 
-def test_initializer_set():
+def test_initializer_set() -> None:
     component_set = InitializerComponentSet()
     for i in range(10):
         component = InitializingComponent(str(i))
@@ -110,7 +114,7 @@ def test_setting_query_with_get_view(
 def test_setting_columns_with_get_view(
     columns: str | list[str], expected_columns: list[str] | None
 ) -> None:
-    expected_columns = expected_columns or columns
+    view_columns = expected_columns or columns
     manager = PopulationManager()
     view = manager._get_view(columns=columns, query="")
-    assert view._columns == expected_columns
+    assert view._columns == view_columns